Move account storage tests to a separate file
